alexaSkill/iot-alexa-skill/requirements.txt
aws-cdk-lib==2.60.0
constructs>=10.0.0,<11.0.0
aws-solutions-constructs.aws-cloudfront-s3==2.35.0
